pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.28.0",
    "langchain>=0.3.0",
    "langchain-community>=0.3.30",
    "langchain-openai>=0.2.0",
    "chromadb>=0.4.0",
    "pypdf2>=3.0.1",
    "python-docx>=1.2.0",
    "python-dotenv>=1.0.0",
    "numpy>=1.24.0",
    "tiktoken>=0.5.0",
    "sentence-transformers>=2.2.0",
]
